refactor(hooks): extract useSketchAnalysis and replace parser useEffect; add unit tests
